% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDiff.R
\name{Comp.Diff}
\alias{Comp.Diff}
\title{Computes differences between two columns/ return: difference}
\usage{
Comp.Diff(db)
}
\arguments{
\item{db}{is a dataframe, with in the first column (1) OTU count and the second column the (average of the) reference frame}
}
\value{
difference between two columns (col1-col2)
}
\description{
Computes differences between two columns/ return: difference
}
