% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPseudoObs.R
\name{Comp.Pseudo.Obs}
\alias{Comp.Pseudo.Obs}
\title{Computes Pseudo Observations for the S-sign transformation and the ratio for the R-sign transformations.}
\usage{
Comp.Pseudo.Obs(db, method)
}
\arguments{
\item{db}{is a dataframe, with in the first column (1) OTU count and the second column the (average of the) reference frame.}

\item{method}{is a string which defines which sign transform is need, SIGN or RANK}
}
\value{
for SIGN transforms is the Pseudo observations (based on the difference) and for RANK transforms is the Ratio
}
\description{
Computes Pseudo Observations for the S-sign transformation and the ratio for the R-sign transformations.
}
