% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompRatio.R
\name{Comp.Ratio}
\alias{Comp.Ratio}
\title{Computes ratio between two columns and avoids inf values/ return: Ratio}
\usage{
Comp.Ratio(db)
}
\arguments{
\item{db}{is a dataframe, with in the first column (1) OTU count and the second column the (average of the) reference frame.
If the reference frame = 0, the ratio is assigned a very large value to avoid inf values. Else the ratio is calculated by dividing the OTU count (col 1) by the (average of the) reference frame (col 2)}
}
\value{
is the ratio (col1 /col2)
}
\description{
Computes ratio between two columns and avoids inf values/ return: Ratio
}
