% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_boot.R
\name{OR_boot}
\alias{OR_boot}
\title{Function to apply the bootstrap method to calculate the standard error of the odds ratio (S-sign RI method)}
\usage{
OR_boot(formula, data, indices)
}
\arguments{
\item{formula}{is the model to be fitted (PO~group + auxiliary)}

\item{data}{is a dataframe with outcome variable PO in first column, group in second, and auxiliary variables}
}
\value{
estimate of the Marginal log Odds ratio
}
\description{
Function to apply the bootstrap method to calculate the standard error of the odds ratio (S-sign RI method)
}
