% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_sign_Conditional.R
\name{R_sign_Conditional}
\alias{R_sign_Conditional}
\title{Probabilistic index model (PIM) for the conditional R-null hypotheses}
\usage{
R_sign_Conditional(formula, data)
}
\arguments{
\item{formula}{is the conditional PIM to fit (Ratio ~ group + auxiliary variables)}

\item{data}{is a dataframe with outcome variable RATIO in first column, grouping variable in the second column and auxiliary variables in the following columns}
}
\value{
list object with 1) fitted PIM 2) estimated PIM coefficients 3) estimated variance-covariance matrix 4) Teststatistic 5) p-value
}
\description{
Probabilistic index model (PIM) for the conditional R-null hypotheses
}
