% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_sign_Marginal.R
\name{R_sign_Marginal}
\alias{R_sign_Marginal}
\title{Probabilistic index model (PIM) for the marginal R-null hypotheses}
\usage{
R_sign_Marginal(data)
}
\arguments{
\item{data}{is a dataframe with outcome variable RATIO in first column and grouping variable in the second column}
}
\value{
list object with 1) fitted PIM 2) estimated PIM coefficients 3) estimated variance-covariance matrix 4) Teststatistic 5) p-value
}
\description{
Probabilistic index model (PIM) for the marginal R-null hypotheses
}
