% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_sign_RI.R
\name{R_sign_RI}
\alias{R_sign_RI}
\title{Regression imputation method for the marginal R-null hypotheses}
\usage{
R_sign_RI(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{is the conditional probabilistic index model to fit (RATIO ~ group + auxiliary variables)}

\item{data}{is a dataframe with outcome variable RATIO in first column, grouping variable in the second column and auxiliary variables in the following columns}
}
\value{
list object with 1) estimated PIM coefficient 2) estimated variance 3) Teststatistic 4) p-value
}
\description{
Regression imputation method for the marginal R-null hypotheses
}
