% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sign.R
\name{S_sign}
\alias{S_sign}
\title{Function to apply S-sign methods}
\usage{
S_sign(formula, Method = "All", db, startdata, ...)
}
\arguments{
\item{formula}{is the model to be fitted}

\item{Method}{can be defined as 'Marginal', 'Conditional', 'RI' or 'All' (default is 'All')}

\item{db}{is a dataframe with raw count of 1 taxon (column 1) and the average reference frame (column 2)}

\item{startdata}{is a phyloseq object with in the sample data only the necessary variables to take into account (first sample variable should be group)}
}
\value{
list of results: S_sign_Marginal & S_sign_Conditional --> 1) logistic regression fit 2) estimated coefficients 3) estimated variance covariance matrix 4) Teststatistic 5) p-value AND S_sign_RI --> 1) estimated coefficients 2) estimated variance covariance matrix 3) Teststatistic 4) p-value
}
\description{
Function to apply S-sign methods
}
