% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sign_Marginal.R
\name{S_sign_Marginal}
\alias{S_sign_Marginal}
\title{Logistic regression model for the marginal S-null hypotheses}
\usage{
S_sign_Marginal(data)
}
\arguments{
\item{data}{is a dataframe with outcome variable PO in first column and grouping variable in the second column}
}
\value{
list object with 1) fitted logistic regression model 2) estimated logistic regression coefficients 3) estimated variance-covariance matrix 4) Teststatistic 5) p-value
}
\description{
Logistic regression model for the marginal S-null hypotheses
}
