% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sign_RI.R
\name{S_sign_RI}
\alias{S_sign_RI}
\title{Regression imputation method for the marginal S-null hypotheses}
\usage{
S_sign_RI(formula, data)
}
\arguments{
\item{formula}{is the conditional logistic regression model to fit (PO ~ group + auxiliary variables)}

\item{data}{is a dataframe with outcome variable PO in first column, grouping variable in the second column and auxiliary variables in the following columns}
}
\value{
list object with 1) estimated logistic regression coefficients 2) estimated variance-covariance matrix 3) Teststatistic 4) p-value
}
\description{
Regression imputation method for the marginal S-null hypotheses
}
