% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_Ref_frame.R
\name{Select_Ref_frame}
\alias{Select_Ref_frame}
\title{Function to select taxa for the reference frame and compute the average}
\usage{
Select_Ref_frame(data, threshold)
}
\arguments{
\item{data}{is a phyloseq object of the raw trimmed counts}

\item{threshold}{to select reference taxa for the reference frame}
}
\value{
a list object with 1) phyloseq object with counts of the non reference taxa 2) vtecor of the average of the reference frame
}
\description{
Function to select taxa for the reference frame and compute the average
}
