% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{Trim}
\alias{Trim}
\title{Trims OTU's that were present in at most 'minReads' sample(s) and with a prevalence less than 'minPrev'
Prevalence is the fraction of samples in which an OTU is observed at least `minReads` times}
\usage{
Trim(obj, minReads = 1L, minPrev = 0.05)
}
\arguments{
\item{obj}{is a phyloseq object or (count) matrix}

\item{minReads}{is the minimum times an OTU was observed over all samples | Default=1}

\item{minPrev}{is the minimum fraction of samples in which an OTU is observed at least `minReads` times | Default=0.05}
}
\value{
trimmed phyloseq object or (count)matrix
}
\description{
Trims OTU's that were present in at most 'minReads' sample(s) and with a prevalence less than 'minPrev'
Prevalence is the fraction of samples in which an OTU is observed at least `minReads` times
}
